fix: rename AtsService to CrmService and update exports
- Renamed lib/AtsService.ts to lib/CrmService.ts (build system requirement)
- Updated lib/index.ts to export CrmService
- Added metadata export to index.ts
- Simplified API handlers
