+++ packages/app-store/lever/index.ts
@@ -1,2 +1,3 @@
 export * as api from "./api";
 export * as lib from "./lib";
+export { metadata } from "./_metadata";
