+++ packages/app-store/lever/lib/index.ts
@@ -1,1 +1,1 @@
-export { default as AtsService } from "./AtsService";
+export { default as CrmService } from "./CrmService";
